requirements.txt.txt
streamlit==1.31.0
langchain==0.1.10
langchain-community==0.0.25
sentence-transformers==2.3.1
chromadb==0.4.22
openai==1.12.0
python-dotenv==1.0.1